update log request

+++ log/src/net_log_request.h
@@ -5,8 +5,8 @@
 #define LOG_PRODUCER_SEND_MAGIC_NUM 0x1B35487A
 
 struct net_log_lz4_buf {
-    size_t length;
-    size_t raw_length;
+    uint32_t length;
+    uint32_t raw_length;
     unsigned char data[0];
 };
